Split models gallery card component
